Refactor release workflow, update documentation links, and enhance code structure
- Adjusted image file copying in release workflow.
- Updated documentation and issue tracker links in manifest.json.
- Refactored image handling in rollup.config.js.
- Improved function calls in classes.js for better clarity.
- moved htmlShutter class to own source file
- moved xyPair class to own source fiel
