+++ src/code/classes.js
@@ -13,6 +13,7 @@ import {
 
 import * as HtmlBlocks from './htmlBlocks.js';
 import {EscImages} from './escImages.js';
+import {xyPair} from './xyPair.js';
 
 
 export class EnhancedShutterCardNew extends LitElement{
@@ -1204,7 +1205,7 @@ export class EnhancedShutter extends LitElement
   actualGlobalWidthPx() {
     let width;
     if (this.actualWidthEdit) {
-      width= this.actualWidthEdit; // Should be solved by an async /await / promise ...
+      width = this.actualWidthEdit; // Should be solved by an async /await / promise ...
     }else{
       width = this[C.ESC_CLASS_SELECTOR]?.getBoundingClientRect()?.width ?? this.cfg.windowWidthPx();
     }
@@ -2013,25 +2014,25 @@ export class shutterCfg {
   }
 
   applyInvertForPositionToText(setting,debug=false){
-    setting = this.applyInvertOpenClose(setting,debug);
+    setting = this.applyInvertOpenCloseUi(setting,debug);
     return setting;
   }
   applyInvertForOverlayDisplay(setting,debug=false){
-    setting = this.applyInvertOpenClose(setting,debug);
+    setting = this.applyInvertOpenCloseUi(setting,debug);
     return setting;
   }
   applyInvertForShowButtonUpDownLabel(setting,debug=false){
-    setting = this.applyInvertOpenClose(setting,debug);
+    setting = this.applyInvertOpenCloseUi(setting,debug);
     setting = this.applyInvertDirection(setting,debug);
     return setting;
   }
   applyInvertForShowButtonUpDownClick(setting,debug){
     setting = this.applyInvertDirection(setting,debug);
-    setting = this.applyInvertCommands(setting,debug);
+    setting = this.applyInvertOpenCloseCover(setting,debug);
     return setting;
   }
   applyInvertForButtonOpenCloseHideStates(setting,debug=false){
-    setting = this.applyInvertOpenClose(setting,debug);
+    setting = this.applyInvertOpenCloseUi(setting,debug);
     setting = this.applyInvertDirection(setting,debug);
     return setting;
   }
@@ -2039,15 +2040,15 @@ export class shutterCfg {
     return setting;
   }
   applyInvertOpenCloseAndPercentage(setting,debug=false){
-    setting = this.applyInvertOpenClose(setting,debug);
-    setting = this.applyInvertPercentage(setting,debug);
+    setting = this.applyInvertOpenCloseUi(setting,debug);
+    setting = this.applyInvertPercentageCover(setting,debug);
     return setting;
   }
   applyInvertAll(setting,debug=false){
-    setting = this.applyInvertOpenClose(setting,debug);
-    setting = this.applyInvertPercentage(setting,debug);
+    setting = this.applyInvertOpenCloseUi(setting,debug);
+    setting = this.applyInvertPercentageCover(setting,debug);
     setting = this.applyInvertDirection(setting,debug);
-    setting = this.applyInvertCommands(setting,debug);
+    setting = this.applyInvertOpenCloseCover(setting,debug);
     return setting;
   }
 
@@ -2056,19 +2057,19 @@ export class shutterCfg {
     return setting;
   }
 
-  applyInvertOpenClose(setting){
+  applyInvertOpenCloseUi(setting){
     if (this.invertOpenCloseUi()) setting = Object.keys(C.INVERT_OPEN_CLOSE_SETTING).includes(setting) ? C.INVERT_OPEN_CLOSE_SETTING[setting] : setting;
     return setting;
   }
-  applyInvertCommands(setting){
+  applyInvertOpenCloseCover(setting){
     if (this.invertOpenCloseCover()) setting = Object.keys(C.INVERT_OPEN_CLOSE_SETTING).includes(setting) ? C.INVERT_OPEN_CLOSE_SETTING[setting] : setting;
     return setting;
   }
-  applyInvertPercentage(setting){
+  applyInvertPercentageCover(setting){
     if (this.invertPercentageCover()) setting = Object.keys(C.INVERT_OPEN_CLOSE_SETTING).includes(setting) ? C.INVERT_OPEN_CLOSE_SETTING[setting] : setting;
     return setting;
   }
-  applyInvertUiPercentage(setting){
+  applyInvertPercentageUi(setting){
     if (this.invertPercentageUi()) setting = Object.keys(C.INVERT_OPEN_CLOSE_SETTING).includes(setting) ? C.INVERT_OPEN_CLOSE_SETTING[setting] : setting;
     return setting;
   }
@@ -2104,11 +2105,11 @@ export class shutterCfg {
         escState= C.SHUTTER_STATE_PARTIAL_OPEN;
       }else{
         // shutter is 0% or 100%
-        escState = position ? this.applyInvertOpenClose(C.SHUTTER_STATE_OPEN) : this.applyInvertOpenClose(C.SHUTTER_STATE_CLOSED);
+        escState = position ? this.applyInvertOpenCloseUi(C.SHUTTER_STATE_OPEN) : this.applyInvertOpenCloseUi(C.SHUTTER_STATE_CLOSED);
       }
     }else  {
       //  shutter is moving,
-      escState = this.applyInvertOpenClose(state);
+      escState = this.applyInvertOpenCloseUi(state);
       //escState = this.applyInvertOpenCloseAndPercentage(state);
 
     }
@@ -2616,151 +2617,5 @@ export class haSubEntity{
   }
   update(haEntity){
     this.entity=haEntity;
-  }
-}
-export class xyPair{
-  #coordX;
-  #coordY;
-  constructor(x=0,y=0){
-    this.#coordX = x;
-    this.#coordY = y;
-  }
-  x(){
-    return this.#coordX;
-  }
-  y(){
-    return this.#coordY;
-  }
-  switch(){
-    const tmp= this.#coordX;
-    this.#coordX = this.#coordY;
-    this.#coordY = tmp;
-  }
-  size(){
-    return this.x()*this.y();
-  }
-  rotate90(){
-    const tmp= this.#coordX;
-    this.#coordX = -this.#coordY;
-    this.#coordY = tmp;
-  }
-  rotate180(){
-    this.#coordX = -this.#coordX;
-    this.#coordY = -this.#coordY;
-  }
-  rotate270(){
-    const tmp= this.#coordX;
-    this.#coordX = this.#coordY;
-    this.#coordY = -tmp;
-  }
-  rotate360(){
-  }
-  fill(x,y){
-    this.#coordX = x;
-    this.#coordY = y;
-  }
-  fill2(xy){
-    this.#coordX = xy.x();
-    this.#coordY = xy.y();
-  }
-}
-export class htmlShutter{
-
-  constructor(enhancedShutter){
-    this.enhancedShutter=enhancedShutter;
-    this.cfg =enhancedShutter.cfg;
-    this.actualScreenPosition = enhancedShutter.actualScreenPosition;
-    this.actualTiltPosition = enhancedShutter.actualTiltPosition;
-    this.actualShutterPosition = enhancedShutter.actualShutterPosition;
-    this.positionText =this.cfg.computePositionText(enhancedShutter.actualShutterPosition,this.actualTiltPosition);
-    this.escImages= enhancedShutter.escImages;
-  }
-
-  defStyleVarsShutter(){
-    let escState=this.cfg.positionToState();
-    const viewImage=this.escImages.getViewImageSrc(this.cfg.id());
-
-    // solves #103 see other lines with shutterSlatImage
-    const shutterSlatImage=this.escImages.getShutterSlatImageSrc(this.cfg.id());
-    const shutterBottomImage=this.escImages.getShutterBottomImageSrc(this.cfg.id());
-
-    return `
-      --mdc-icon-button-size: ${this.cfg.iconButtonSize()}${C.UNITY};
-      --ha-icon-button-size: ${this.cfg.iconButtonSize()}${C.UNITY};
-      --mdc-icon-size: ${this.cfg.iconSize()}${C.UNITY};
-      --esc-icon-size-wifi-battery: ${this.cfg.iconSizeWifiBattery()}${C.UNITY};
-      --esc-icon-div-size: ${C.ICON_DIV_SIZE/C.ICON_SIZE*this.cfg.iconSizeWifiBattery()}${C.UNITY};
-      --esc-icons-margins: ${this.cfg.iconsPosition()==C.TOP
-          ? `${C.ICON_MARGIN_TB}${C.UNITY} ${C.ICON_MARGIN_LR}${C.UNITY} auto ${C.ICON_MARGIN_LR}${C.UNITY}`
-          : `auto ${C.ICON_MARGIN_LR}${C.UNITY} ${C.ICON_MARGIN_TB}${C.UNITY} ${C.ICON_MARGIN_LR}${C.UNITY}`};
-
-      --esc-overflow: ${this.enhancedShutter.getOverflow()};
-
-      --esc-flex-name_opening-flow: ${this.cfg.inlineHeader() ? 'row' : 'column'} nowrap;
-      --esc-flex-flow-middle: ${!this.cfg.buttonGroupInRow() ? 'column': 'row'}${this.cfg.buttonsContainerReversed() ? '-reverse' : ''} nowrap;
-      --esc-window-height: ${this.cfg.windowHeightPx()+C.UNITY};
-      --esc-window-width1: ${this.cfg.buttonGroupInRow() ? '100%': this.cfg.windowWidthPx()+C.UNITY};
-      --esc-window-width: ${this.cfg.windowWidthPx()+C.UNITY};
-      --esc-window-background-image: ${viewImage.includes('.') ?  `url(${viewImage})` : ''};
-      --esc-window-background-color: ${viewImage.includes('.') ? '' : `${viewImage}`};
-      --esc-window-rotate: ${this.cfg.viewImageRotate()};
-      --esc-button-rotate: ${this.cfg.buttonRotate()};
-
-      --esc-transform-slide:  ${this.enhancedShutter.transformSlide(this.actualScreenPosition)};
-      --esc-transform-picker: ${this.enhancedShutter.transformPicker(this.actualScreenPosition)};
-      --esc-tilt-angle-deg: ${this.enhancedShutter.getTiltAngleDeg(this.enhancedShutter.react_TiltPosition)};
-      --esc-tilt-angle-deg-graph: ${this.enhancedShutter.getTiltAngleDegGraph(this.enhancedShutter.react_TiltPosition)};
-
-      --esc-transform-undo-slats-rotate:  ${this.enhancedShutter.transformUndoSlatsRotate()};
-      --esc-transform-tilt-slat-rotate:  ${this.enhancedShutter.transformTiltSlatRotate()};
-      --esc-transform-movement: ${this.enhancedShutter.transformMovement()};
-
-      --esc-picker-top: -${this.cfg.pickerOverlapPx()+C.UNITY};
-      --esc-picker-height: ${this.cfg.pickerOverlapPx()*2+C.UNITY};
-
-      --esc-slat-height: ${this.enhancedShutter.slatHeightPx()+C.UNITY};
-
-      --esc-tilt-slat-height: ${this.enhancedShutter.tiltSlatHeightPx()+C.UNITY};
-      --esc-tilt-slat-width: ${this.enhancedShutter.tiltSlatWidthPx()};
-      --esc-tilt-slat-origin: ${this.enhancedShutter.tiltSlatOrigin()};
-      --esc-tilt-slat-background-size: ${this.enhancedShutter.tiltSlatBackgroundSize()};
-      --esc-slider-writing-mode: ${this.enhancedShutter.sliderWritingMode()};
-      --esc-slider-direction: ${this.enhancedShutter.sliderDirection()};
-      --esc-tilt-icon-rotate: ${(this.enhancedShutter.tiltIconRotate3())};
-
-      --esc-slide-slats-height: ${this.enhancedShutter.slatsSlideHeightPx()+C.UNITY};
-      --esc-slide-edge-height: ${this.enhancedShutter.shutterBottomSize().y()+C.UNITY};
-
-      --esc-transform-partial: ${this.enhancedShutter.transformPartial()};
-
-      --esc-buttons-flex-flow:      ${!this.cfg.buttonGroupInRow() ? 'row-reverse' : 'column'} nowrap;
-      --esc-buttons-flex-flow-tilt: ${!this.cfg.buttonGroupInRow() ? 'row-reverse' : 'column'} nowrap;
-
-      --esc-movement-overlay-display: ${(escState == C.SHUTTER_STATE_OPENING || escState == C.SHUTTER_STATE_CLOSING) ? 'block' : C.NONE};
-      --esc-movement-overlay-up-display: ${escState == this.cfg.applyInvertForOverlayDisplay(C.SHUTTER_STATE_OPENING) ? 'block' : C.NONE};
-      --esc-movement-overlay-down-display: ${escState == this.cfg.applyInvertForOverlayDisplay(C.SHUTTER_STATE_CLOSING) ? 'block' : C.NONE};
-
-      --esc-slide-background-main-image: ${shutterSlatImage.includes('.') ?  `url(${shutterSlatImage})` : ''};
-      --esc-slide-background-edge-image: ${shutterBottomImage.includes('.') ?  `url(${shutterBottomImage})` : ''};
-
-      --esc-slide-background-main-color: ${shutterSlatImage.includes('.') ? '' : `${shutterSlatImage}`};
-      --esc-slide-background-edge-color: ${shutterBottomImage.includes('.') ? '' : `${shutterBottomImage}`};
-
-      --esc-slide-background-slat-size: ${this.enhancedShutter.shutterSlatSizePercentage()};
-      --esc-slide-background-slats-size: ${this.enhancedShutter.shutterSlatsSizePercentage()};
-      --esc-slide-background-edge-size: ${this.enhancedShutter.shutterBottomSizePercentage()};
-
-      --esc-slide-background-main-position: ${this.enhancedShutter.shutterMainBackgroundPosition()};
-      --esc-slide-background-edge-position: ${this.enhancedShutter.shutterEdgeBackgroundPosition()};
-
-      --esc-top-right-color: ${this.cfg.signalIconColor()};
-      --esc-top-left-color: ${this.cfg.batteryIconColor()};
-
-      --esc-top-icon-text-line-height: ${this.cfg.iconScalePercent()};
-      --esc-top-icon-text-font-size: ${this.cfg.iconScalePercent()};
-      --esc-text-scale: ${this.cfg.textScaleFactor()};
-      --esc-button-scale: ${this.cfg.buttonScaleFactor()};
-
-    `;
   }
 }
